change:move the main code to src
