+++ src/misaka.py
@@ -1,7 +1,7 @@
 import sys
 
 from PyQt5.QtWidgets import QApplication, QMainWindow
-import ui
+from src import ui
 import requests
 import time
